move SO3 to so3.py, create special_orthogonal.py for SOn
