+++ examples/plot_geodesics_so3.py
@@ -10,7 +10,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 import geomstats.visualization as visualization
-from geomstats.geometry.special_orthogonal import SpecialOrthogonal
+from geomstats.geometry.so3 import SpecialOrthogonal
 
 SO3_GROUP = SpecialOrthogonal(n=3)
 METRIC = SO3_GROUP.bi_invariant_metric
