+++ tests/test_special_orthogonal.py
@@ -9,7 +9,7 @@ import tests.helper as helper
 import geomstats.backend as gs
 import geomstats.tests
 from geomstats.geometry.invariant_metric import InvariantMetric
-from geomstats.geometry.special_orthogonal import SpecialOrthogonal
+from geomstats.geometry.so3 import SpecialOrthogonal
 
 
 EPSILON = 1e-5
